Conflicts:
	P2P/library/api/fermat-p2p-api/src/main/java/com/bitdubai/fermat_p2p_api/layer/all_definition/communication/network_services/data_base/CommunicationNetworkServiceDeveloperDatabaseFactory.java

+++ BNK/plugin/wallet/fermat-bnk-plugin-wallet-bank-money-bitdubai/src/main/java/com/bitdubai/fermat_bnk_plugin/layer/wallet/bank_money/developer/bitdubai/version_1/structure/BankMoneyWalletBalanceImpl.java
@@ -8,6 +8,8 @@ import com.bitdubai.fermat_bnk_plugin.layer.wallet.bank_money.developer.bitdubai
 import com.bitdubai.fermat_bnk_plugin.layer.wallet.bank_money.developer.bitdubai.version_1.exceptions.CantAddCreditException;
 import com.bitdubai.fermat_bnk_plugin.layer.wallet.bank_money.developer.bitdubai.version_1.exceptions.CantAddDebitException;
 
+import java.math.BigDecimal;
+
 /**
  * Created by memo on 23/11/15.
  */
@@ -23,7 +25,7 @@ public class BankMoneyWalletBalanceImpl implements BankMoneyWalletBalance{
     }
 
     @Override
-    public double getBalance(String accountNumber) throws CantCalculateBalanceException {
+    public BigDecimal getBalance(String accountNumber) throws CantCalculateBalanceException {
         switch (balanceType){
             case AVAILABLE: return bankMoneyWalletDao.getAvailableBalance(accountNumber);
             case BOOK: return bankMoneyWalletDao.getBookBalance(accountNumber);
